+++ diffractio/vector_sources_XY.py
@@ -399,11 +399,6 @@ class Vector_source_XY(Vector_field_XY):
     #     """Pass Ex, Ey field to py_pol package for software analysis
     #     """
 
-    #     j0 = Jones_vector(name="from Diffractio")
-    #     j0.from_components(Ex=self.Ex, Ey=self.Ey)
-
-    #     return j0
-
 
 def define_initial_field(EM, u=None):
     """Defines the initial field EM = (Ex, Ey) in terms of u.